refactor: organize api routes and helpers
